+++ test/support/ecto_fake_app/repo.ex
@@ -3,21 +3,12 @@ defmodule Permit.EctoFakeApp.Repo do
     otp_app: :permit_phoenix,
     adapter: Ecto.Adapters.Postgres
 
-  alias Permit.EctoFakeApp.{Item, Repo, User}
+  alias Permit.EctoFakeApp.Repo
+  alias Permit.EctoFakeApp.SeedData
 
   def seed_data! do
-    users = [
-      %User{id: 1} |> Repo.insert!(),
-      %User{id: 2} |> Repo.insert!(),
-      %User{id: 3} |> Repo.insert!()
-    ]
-
-    items = [
-      %Item{id: 1, owner_id: 1, permission_level: 1} |> Repo.insert!(),
-      %Item{id: 2, owner_id: 2, permission_level: 2, thread_name: "dmt"} |> Repo.insert!(),
-      %Item{id: 3, owner_id: 3, permission_level: 3} |> Repo.insert!()
-    ]
-
+    users = SeedData.users() |> Enum.map(&Repo.insert!(&1))
+    items = SeedData.items() |> Enum.map(&Repo.insert!(&1))
     %{users: users, items: items}
   end
 end